spyro-agentic-rag/requirements.txt
# Core dependencies
neo4j-graphrag[openai]==1.9.0
langchain==0.3.18
langchain-openai==0.2.14
langchain-community==0.3.17

# API framework
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.11.0
python-multipart==0.0.9

# Neo4j driver
neo4j==5.28.0

# Environment and utilities
python-dotenv==1.0.0
httpx==0.28.0

# Logging and monitoring
structlog==24.4.0
rich==13.7.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Development
black==24.4.2
ruff==0.5.0
mypy==1.10.0
pre-commit==3.7.1